Make selection work for MVT layers
It's a bit rudimentary as elements can't be unselected and end up in the selection array multiple times (when clicked multiple times).

Also added a `convertStyleToOpenLayers` function that takes the object we've been using so far (for Leaflet and in the workspace descriptions) and turns that into OpenLayers's format which uses lots of classes: an `ol.Style` constructed with `ol.Stroke` and `ol.Fill`.

Also amended the `hexToRGBA` function for that so it returns its result as an integer array (instead of a string) if passed an optional parameter.