Make tile URL templates with `{s}` placeholder work
In Leaflet, `{s}` is a placeholder for the possible subdomains, like in `https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png`. It is configurable, but defaults to choose one of `a`, `b`, `c`. In OpenLayers, the range has to be specified directly in the template, so the default `{s}` becomes `{a-c}`. We should therefore replace the former with the latter for backwards compatibility.