Set default minZoom=2, derive view from bounds if not specified
With minZoom=0 you get to see the entire world about three times; minZoom=2 is more "see the whole globe once".

View is not mandatory to be specified, but bounds are always there, so we can derive an appropriate view from the bounds (just average the coordinates). The zoom is derived from the minZoom so it's always as far out as is allowed.