Fix events (rendercomplete, click for mvt selection)
The `rendercomplete` and `render` events are in fact fired by the map, not by the individual layers -- this is wrong in OL's docs, see openlayers/openlayers#9168.

The `click` event was added to the map twice: First, as it should, by the `add -> loadLocale -> init` call. But then *again* by the mvt's `viewchangeend` event handler function, because that event is fired before the first `rendercomplete` is fired, so that `layer.loaded` isn't set to  `true` just yet. This causes the "stop" at the beginning of `mvt.mjs` to not take effect and thus the code to accidentally run through all the way, adding the event handler a second time. This second event handler reversed the changes done to the `layer.selected` set by the first one, causing selecting items on the map to be uneffective.