Make cluster layers work in OL (basic support)
Instead of passing a GeoJSON layer, we can convert the incoming data into Point features directly. In OL we don't need a `pointToLayer` function, but have to set the style for the Point feature. Only `categorized` theme tested so far, but that works quite good, even with the multiple discs to create the "supermarket logo color markers" :-)