Display cursor as pointer when hovering clickable feature
Also fixes a bug where features were highlighted although they were not directly below the cursor and thus in fact not clickable (e.g. a when a popup is open and covers parts of a clickable feature, the feature was highlighted when hovering the popup, although an eventual click wouldn't target the feature but the popup).