Enable locate feature
In Open Layers, this is a bit more complicated: You can't just add a marker to a map, but must create a point feature, style that as a marker, supply this to a VectorSource, which is then displayed on its own VectorLayer.

Minor bug: On a cluster layer, the geolocation marker ends up below the cluster layer's content after moving the map, and pops back to the top once the geolocation watcher reports new coordinates. I set the marker layer's zIndex to 100 (whereas the cluster should have 40 at maximum), but that didn't make it go away.